+++ tests/integration/test_projectconfigs.py
@@ -201,3 +201,53 @@ def test_pending_projects(mini_sentry, relay):
     print(data)
     assert public_key in data["configs"]
     assert data.get("pending") is None
+
+
+def test_fallback_to_v2(mini_sentry, relay):
+    mini_sentry.project_config_simulate_pending = True
+    relay = relay(
+        mini_sentry, wait_healthcheck=True, options={"limits": {"query_timeout": 1}}
+    )
+    project = mini_sentry.add_basic_project_config(42)
+    public_key = mini_sentry.get_dsn_public_key(42)
+
+    body = {"publicKeys": [public_key]}
+    packed, signature = SecretKey.parse(relay.secret_key).pack(body)
+
+    def request_config():
+        return relay.post(
+            "/api/0/relays/projectconfigs/?version=3",
+            data=packed,
+            headers={
+                "X-Sentry-Relay-Id": relay.relay_id,
+                "X-Sentry-Relay-Signature": signature,
+            },
+        )
+
+    # Trigger a config request
+    response = request_config()
+    assert response.ok
+    data = response.json()
+    assert public_key in data["pending"]
+    assert public_key not in data["configs"]
+
+    # Give relay time to query sentry:
+    time.sleep(1)
+
+    # Error was logged:
+    _, last_error = mini_sentry.test_failures[-1]
+    assert "Failed to fetch project config from V3 endpoint" in str(last_error)
+    mini_sentry.test_failures.clear()
+
+    # Relay attempted to query version 3, but then fell back to version 2:
+    assert "?version=3" in mini_sentry.request_log[-3]
+    assert "?version=2" in mini_sentry.request_log[-2]
+
+    assert "/envelope/" in mini_sentry.request_log[-1]  # The error that was logged
+
+    # Verify that valid config is now available:
+    response = request_config()
+    assert response.ok
+    data = response.json()
+    assert public_key in data["configs"]
+    assert data.get("pending") is None